+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Type errors are detected even in files with parse errors.
 - The `--no-debug` flag has been renamed to `--no-debug-ops`.
 - The `jump` debugging command has been renamed to `goto`.
+- When `__eval` results in a Python exception, the error message with the location is printed to standard error, rather than the raw Python traceback.
 
 ### Removed
 - The `jump` debugging command (now called `goto`) may no longer be invoked with no arguments.